hanko/requirements.txt
kubernetes>=24.2.0
hvac>=1.2.1
pyyaml>=6.0.1
gitpython>=3.1.43